cmake changes for building on VS/NVIDIA
1. CMakeLists.txt - added GMX_USE_OPENCL option
2. gmxDetectGpu.cmake - added gmx_find_OpenCL macro. To replace
gmx_detect_OpenCL?
It searches for OpenCL include and lib folders.
It should be cross platform and work for AMD, NVIDIA, Intel.
This is an adapted version of the code here:

https://gitorious.org/findopencl/chuckatkinss-findopencl/source/695c5f6a754baf63d8b5c79a40c35d86e47a140a:FindOpenCL.cmake
3. gmxManageOpenCL.make - replaced gmx_detect_OpenCL call by
gmx_find_OpenCL call
Removed include and lib folders initializations as they are already
implemented in gmx_detect_OpenCL