Ran Gromacs code beautifier
If needed in future, this ran with

admin/uncrustify.sh --rev=v5.0-alpha.0 update-workdir

but (unfortunately) we need to use a custom fork of uncrustify
(details noted in that script).

Passing later Gromacs review needs this script to be happy, so
we should start now in case it makes merges easier too!